Address PR comments
1. switch to using uint64_t
2. overload putbytestring
3. define max event length for process sample
4. and normalize cpu user and sys time

+++ src/os.h
@@ -10,6 +10,7 @@
 #include <stddef.h>
 #include <sys/types.h>
 #include "arch.h"
+#include <cstdint>
 
 
 typedef void (*SigAction)(int, siginfo_t*, void*);
@@ -32,30 +33,29 @@ struct ProcessInfo {
     char _cmdline[2048];         // Command line from /proc/{pid}/cmdline
     unsigned int _uid;           // User ID
     unsigned char _state;        // Process state (R, S, D, Z, T, etc.)
-    unsigned long _start_time;   // Process start time (clock ticks since boot)
+    uint64_t _start_time;        // Process start time (clock ticks since boot)
 
     // CPU & thread stats
-    unsigned long _cpu_user;     // User CPU time (clock ticks)
-    unsigned long _cpu_system;   // System CPU time (clock ticks)
+    uint64_t _cpu_user;          // User CPU time (seconds)
+    uint64_t _cpu_system;        // System CPU time (seconds)
     float _cpu_percent;          // CPU utilization percentage
-    unsigned short _threads;     // Number of threads
+    uint64_t _threads;           // Number of threads
 
     // Memory stats (in KB)
-    unsigned long _vm_size;     // Total virtual memory size
-    unsigned long _vm_rss;      // Resident memory size
-    unsigned long _rss_anon;     // Resident anonymous memory
-    unsigned long _rss_files;    // Resident file mappings
-    unsigned long _rss_shmem;    // Resident shared memory
+    uint64_t _vm_size;           // Total virtual memory size
+    uint64_t _vm_rss;            // Resident memory size
+    uint64_t _rss_anon;          // Resident anonymous memory
+    uint64_t _rss_files;         // Resident file mappings
+    uint64_t _rss_shmem;         // Resident shared memory
 
     // Page fault stats
-    unsigned long _minor_faults; // Minor page faults (no I/O required)
-    unsigned long _major_faults; // Major page faults (I/O required)
+    uint64_t _minor_faults;      // Minor page faults (no I/O required)
+    uint64_t _major_faults;      // Major page faults (I/O required)
 
     // I/O stats
-    long _io_read;      // KB read from storage
-    long _io_write;     // KB written to storage
-
-    unsigned long _last_update;  // Timestamp of last update
+    uint64_t _io_read;           // KB read from storage
+    uint64_t _io_write;          // KB written to storage
+    uint64_t _last_update;       // Timestamp of last update
 
     ProcessInfo() : _pid(0), _ppid(0), _uid(0), _state(0), _start_time(0),
                     _cpu_user(0), _cpu_system(0), _cpu_percent(0.0F), _threads(0),
@@ -107,7 +107,7 @@ class OS {
   public:
     static const size_t page_size;
     static const size_t page_mask;
-    static const long clock_ticks_per_sec;  // System clock ticks per second (_SC_CLK_TCK)
+    static const long clock_ticks_per_sec;
 
     static u64 nanotime();
     static u64 micros();
